requirements.txt
fastapi
uvicorn
neo4j
numpy
langchain
openai
pydantic
psycopg2-binary
sqlalchemy
